+++ packages/react-native/package.json
@@ -20,6 +20,11 @@
       "react-native": "./src/adapters/executorch.ts",
       "import": "./dist/adapters/executorch.js"
     },
+    "./adapters/instant-cv": {
+      "types": "./dist/adapters/instant-cv.d.ts",
+      "react-native": "./src/adapters/instant-cv.ts",
+      "import": "./dist/adapters/instant-cv.js"
+    },
     "./adapters/vision-camera": {
       "types": "./dist/adapters/vision-camera.d.ts",
       "react-native": "./src/adapters/vision-camera.ts",
@@ -29,6 +34,11 @@
       "types": "./dist/adapters/video-file.d.ts",
       "react-native": "./src/adapters/video-file.ts",
       "import": "./dist/adapters/video-file.js"
+    },
+    "./worklet-debug": {
+      "types": "./dist/worklet-debug.d.ts",
+      "react-native": "./src/worklet-debug.ts",
+      "import": "./dist/worklet-debug.js"
     },
     "./media-session": {
       "types": "./dist/media-session.d.ts",
